requirements.txt
# langchain
# langchain-core

# langchain-openai
# langchain-anthropic
# langchain-google-genai
# langchain-huggingface

# google-generativeai
# transformers
# huggingface-hub
# python-dotenv
# numpy
# scikit-learn
# ipython

annotated-types==0.7.0
anthropic==0.81.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
filelock==3.24.2
filetype==1.2.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-core==1.2.13
langchain-google-genai==4.2.0
langchain-huggingface==1.2.0
langchain-openai==1.1.10
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.4
matplotlib-inline==0.2.1
numpy==2.4.2
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parso==0.8.6
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.6.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
